+++ src/TestSuite/TestTrait.php
@@ -18,7 +18,6 @@ namespace Tools\TestSuite;
 
 use BadMethodCallException;
 use Exception;
-use PHPUnit\Framework\Error\Error;
 use Throwable;
 
 /**
@@ -123,8 +122,6 @@ trait TestTrait
 
         try {
             call_user_func($function);
-        } catch (Error $e) {
-            throw $e;
         } catch (Exception $e) {
             parent::assertInstanceof(
                 $expectedException,