+++ src/or_fail_functions.php
@@ -22,24 +22,25 @@ use Tools\Exception\NotPositiveException;
 use Tools\Exception\NotReadableException;
 use Tools\Exception\NotWritableException;
 use Tools\Exception\PropertyNotExistsException;
+use Tools\Exceptionist;
 
 if (!function_exists('file_exists_or_fail')) {
     /**
-     * Checks whether a file or directory exists and throws an exception if the
-     *  file does not exist
+     * Checks whether a file or directory exists and throws an exception if does
+     *  not exist
      * @param string $filename Path to the file or directory
      * @param string|null $message The failure message that will be appended to the
      *  generated message
      * @param \Throwable|string $exception The exception class you want to set
      * @return string
+     * @deprecated Use `Exceptionist::fileExists()` instead
      * @throws \Tools\Exception\FileNotExistsException
      */
     function file_exists_or_fail(string $filename, $message = null, $exception = FileNotExistsException::class): string
     {
-        $message = $message ?: sprintf('File or directory `%s` does not exist', rtr($filename));
-        is_true_or_fail(is_readable($filename), $message, $exception);
+        deprecationWarning('Deprecated. Use `Exceptionist::fileExists()` instead.');
 
-        return $filename;
+        return Exceptionist::fileExists($filename, $message, $exception);
     }
 }
 
@@ -53,14 +54,17 @@ if (!function_exists('in_array_or_fail')) {
      *  the generated message
      * @param \Throwable|string $exception The exception class you want to set
      * @return mixed
+     * @deprecated Use `Exceptionist::inArray()` instead
      * @since 1.2.6
      */
     function in_array_or_fail($value, array $array, $message = null, $exception = NotInArrayException::class)
     {
+        deprecationWarning('Deprecated. Use `Exceptionist::inArray()` instead.');
+
         if (!$message && is_stringable($value) == 'The value is not in array') {
             $message = sprintf('The value `%s` is not in array', (string)$value);
         }
-        is_true_or_fail(in_array($value, $array), $message, $exception);
+        Exceptionist::inArray([$value, $array], $message, $exception);
 
         return $value;
     }
@@ -68,63 +72,69 @@ if (!function_exists('in_array_or_fail')) {
 
 if (!function_exists('is_dir_or_fail')) {
     /**
-     * Tells whether the filename is a directory and throws an exception if the
+     * Checks whether the filename is a directory and throws an exception if the
      *  filename is not a directory
      * @param string $filename Path to the directory
      * @param string|null $message The failure message that will be appended to the
      *  generated message
      * @param \Throwable|string $exception The exception class you want to set
      * @return string
+     * @deprecated Use `Exceptionist::isDir()` instead
      * @throws \Tools\Exception\NotDirectoryException
      */
     function is_dir_or_fail(string $filename, $message = null, $exception = NotDirectoryException::class): string
     {
-        $message = $message ?: sprintf('Filename `%s` is not a directory', rtr($filename));
-        is_true_or_fail(is_dir($filename), $message, $exception);
+        deprecationWarning('Deprecated. Use `Exceptionist::isDir()` instead.');
 
-        return $filename;
+        $message = $message ?: sprintf('Filename `%s` is not a directory', rtr($filename));
+
+        return Exceptionist::isDir($filename, $message, $exception);
     }
 }
 
 if (!function_exists('is_positive_or_fail')) {
     /**
-     * Throws an exception if the value is not a positive
+     * Checks whether the value is a positive and throws an exception if the
+     *  value is not a positive
      * @param mixed $value The value you want to check
      * @param string $message The failure message that will be appended to the
      *  generated message
      * @param \Throwable|string $exception The exception class you want to set
      * @return mixed
+     * @deprecated Use `Exceptionist::isPositive()` instead
      * @since 1.2.5
      * @throws \Tools\Exception\NotPositiveException
      */
     function is_positive_or_fail($value, $message = null, $exception = NotPositiveException::class)
     {
+        deprecationWarning('Deprecated. Use `Exceptionist::isPositive()` instead.');
+
         if (!$message && is_stringable($value)) {
             $message = sprintf('The value `%s` is not a positive', (string)$value);
         }
-        is_true_or_fail(is_positive($value), $message, $exception);
 
-        return $value;
+        return Exceptionist::isPositive($value, $message, $exception);
     }
 }
 
 if (!function_exists('is_readable_or_fail')) {
     /**
-     * Tells whether a file exists and is readable and throws an exception if
-     *  the file is not readable
+     * Checks whether a file or directory exists and is readable and throws an
+     *  exception if is not readable
      * @param string $filename Path to the file or directory
      * @param string|null $message The failure message that will be appended to
      *  the generated message
      * @param \Throwable|string $exception The exception class you want to set
      * @return string
+     * @deprecated Use `Exceptionist::isReadable()` instead
+     * @throws \Tools\Exception\FileNotExistsException
      * @throws \Tools\Exception\NotReadableException
      */
     function is_readable_or_fail(string $filename, $message = null, $exception = NotReadableException::class): string
     {
-        $message = $message ?: sprintf('File or directory `%s` is not readable', rtr($filename));
-        is_true_or_fail(is_readable($filename), $message, $exception);
+        deprecationWarning('Deprecated. Use `Exceptionist::isReadable()` instead.');
 
-        return $filename;
+        return Exceptionist::isReadable($filename, $message, $exception);
     }
 }
 
@@ -139,54 +149,36 @@ if (!function_exists('is_true_or_fail')) {
      *  generated message
      * @param \Throwable|string $exception The exception class you want to set
      * @return mixed
+     * @deprecated Use `Exceptionist::isTrue()` instead
      * @since 1.1.7
      * @throws \Exception
      */
     function is_true_or_fail($value, $message = 'The value is not equal to `true`', $exception = \ErrorException::class)
     {
-        if ($value) {
-            return $value;
-        }
+        deprecationWarning('Deprecated. Use `Exceptionist::isTrue()` instead.');
 
-        if ($message instanceof \Throwable || (is_string($message) && class_exists($message))) {
-            [$exception, $message] = [$message, 'The value is not equal to `true`'];
-        }
-
-        if (!$exception instanceof \Throwable) {
-            if (!is_string($exception)) {
-                trigger_error('`$exception` parameter must be a string');
-            }
-            if (!class_exists($exception)) {
-                trigger_error(sprintf('Class `%s` does not exist', $exception));
-            }
-            $exception = new $exception($message);
-        }
-
-        if (!$exception instanceof \Throwable) {
-            trigger_error(sprintf('`%s` is not and instance of `Throwable`', get_class($exception)));
-        }
-
-        throw $exception;
+        return Exceptionist::isTrue($value, $message, $exception);
     }
 }
 
 if (!function_exists('is_writable_or_fail')) {
     /**
-     * Tells whether the filename is writable and throws an exception if the
-     *  file is not writable
+     * Checks whether a file or directory exists and is writable and throws an
+     *  exception if is not writable
      * @param string $filename Path to the file or directory
      * @param string|null $message The failure message that will be appended to
      *  the generated message
      * @param \Throwable|string $exception The exception class you want to set
      * @return string
+     * @deprecated Use `Exceptionist::isWritable()` instead
+     * @throws \Tools\Exception\FileNotExistsException
      * @throws \Tools\Exception\NotWritableException
      */
     function is_writable_or_fail(string $filename, $message = null, $exception = NotWritableException::class): string
     {
-        $message = $message ?: sprintf('File or directory `%s` is not writable', rtr($filename));
-        is_true_or_fail(is_writable($filename), $message, $exception);
+        deprecationWarning('Deprecated. Use `Exceptionist::isWritable()` instead.');
 
-        return $filename;
+        return Exceptionist::isWritable($filename, $message, $exception);
     }
 }
 
@@ -202,20 +194,14 @@ if (!function_exists('key_exists_or_fail')) {
      *  the generated message
      * @param string $exception The exception class you want to set
      * @return mixed
+     * @deprecated Use `Exceptionist::arrayKeyExists()` instead
      * @throws \Tools\Exception\KeyNotExistsException
      */
     function key_exists_or_fail($key, array $array, $message = null, $exception = KeyNotExistsException::class)
     {
-        $messageWasEmpty = empty($message);
-        foreach ((array)$key as $name) {
-            $result = array_key_exists($name, $array);
-            if ($messageWasEmpty) {
-                $message = sprintf('Key `%s` does not exist', $name);
-            }
-            is_true_or_fail($result, $message, $exception);
-        }
+        deprecationWarning('Deprecated. Use `Exceptionist::arrayKeyExists()` instead.');
 
-        return $key;
+        return Exceptionist::arrayKeyExists($key, $array, $message, $exception);
     }
 }
 
@@ -232,20 +218,14 @@ if (!function_exists('property_exists_or_fail')) {
      *  the generated message
      * @param \Throwable|string $exception The exception class you want to set
      * @return mixed
+     * @deprecated Use `Exceptionist::objectPropertyExists()` instead
      * @since 1.1.14
      * @throws \Tools\Exception\PropertyNotExistsException
      */
     function property_exists_or_fail($object, $property, $message = null, $exception = PropertyNotExistsException::class)
     {
-        $messageWasEmpty = empty($message);
-        foreach ((array)$property as $name) {
-            if ($messageWasEmpty) {
-                $message = sprintf('Object does not have `%s` property', $name);
-            }
-            $result = method_exists($object, 'has') ? $object->has($name) : property_exists($object, $name);
-            is_true_or_fail($result, $message, $exception);
-        }
+        deprecationWarning('Deprecated. Use `Exceptionist::objectPropertyExists()` instead.');
 
-        return $property;
+        return Exceptionist::objectPropertyExists($object, $property, $message, $exception);
     }
 }